backend/pyproject.toml
[project]
name = "uscornie-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "psycopg2-binary",
    "python-jose[cryptography]",
    "python-multipart",
    "google-auth",
    "requests",
    "pydantic-settings",
]

[dependency-groups]
dev = [
    "ruff",
    "ty",
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "N", "UP", "B", "SIM", "ASYNC", "FAST"
]
ignore = ["E501", "E712", "B008", "FAST002"]

[tool.ruff.format]
quote-style = "double"

[tool.ty]
python-version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
